% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{checkbox}
\alias{checkbox}
\title{Checkbox or Multiple choice Tabulation}
\usage{
  ctoimport(
    column,
    decimal = 1,
  )
}
\arguments{
  \item{column}{A multiple choice or checkbox column}
  
  \item{decimal}{Number of decimal point for the percentages}
}
\value{
  
}
\description{
  Tabulation of checkbox or multiple choice columns
}
\examples{
  NOT RUN
  checkbox(column,decimal)
}
