% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctoimport.R
\name{ctoimport}
\alias{ctoimport}
\title{Importing Data from SurveyCTO server}
\usage{
ctoimport(servername, formid, username, password)
}
\arguments{
\item{servername}{SurveyCTO servername}

\item{formid}{Get this from Setting sheet in the XLSForm workbook}

\item{username}{SurveyCTO username}

\item{password}{SurveyCTO password}
}
\value{

}
\description{
Importing Data from SurveyCTO server
}
\examples{
ctoimport('servername','formid','username','password')
}
