% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_tag.R
\name{dup_tag}
\alias{dup_tag}
\title{Evaluating Duplicates in a column (Indexing or Counting)}
\usage{
  dup_tag(
    column,
    index=FALSE,
    na.rm = TRUE,
  )
}
\arguments{
  \item{column}{the column where ducplicated are beign evaluated }
  
  \item{index}{FALSE is one wants the duplicates to be indixed, and TRUE 
               for duplicated to be indicated by number of occurence.}
  
    \item{na.rm}{TRUE is you want NA or empty responses to be evaluated...}
}
\value{
  
}
\description{
  Evaluating Duplicates in a column (Indexing or Counting)
}
\examples{
  ##NOT RUN
  ##dup_tag(column)
}
