% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboimport.R
\name{koboimport}
\alias{koboimport}
\title{Importing data from Kobo tool box (even Humanitarian Account)}
\usage{
koboimport(servername, formid, username, password)
}
\arguments{
\item{servername}{either 'https://kobo.humanitarianresponse.info/' or 'kc.kobotoolbox.org'}

\item{formid}{Click the form and copy forms/"formid"/landing}

\item{username}{your username}

\item{password}{you password}
}
\value{

}
\description{
Importing data from Kobo tool box (even Humanitarian Account)
}
\examples{
data<-koboimport(servername,form_id,username,password)

}
