% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkmean.R
\name{odkmean}
\alias{odkmean}
\title{Calculates the mean by group}
\usage{
odkmean(row, column, data)
}
\arguments{
\item{row}{Numeric variable}

\item{column}{Group variable}

\item{data}{data where with row and column variables}
}
\value{

}
\description{
Calculates the mean by group
}
\examples{
 odkmean('row','column','data')
}
