% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodkmult.R
\name{rodkmult}
\alias{rodkmult}
\title{Compiling Multiple Variable Choices.}
\usage{
rodkmult(xlsform, dataName)
}
\arguments{
\item{xlsform}{Latest XLSFORM used to collect data.The variable and value
labels will be used to label the data.}

\item{dataName}{The name of the data in the memory collected using the
survey.}
}
\value{

}
\description{
Compiling Multiple Variable Choices.
}
