% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodkmultlab.R
\name{rodkmultlab}
\alias{rodkmultlab}
\title{Compile Multiple Select variable value labels into a data.frame.}
\usage{
rodkmultlab(xlsform)
}
\arguments{
\item{xlsform}{Lasted XLSFORM used to collect the data.}
}
\value{

}
\description{
Compile Multiple Select variable value labels into a data.frame.
}
\examples{
NOT run
rodkmultlab(xlsforn,'dataName')
}
