% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodksinglelab.R
\name{rodksinglelab}
\alias{rodksinglelab}
\title{Compile Single Select variable value label into data.frame.}
\usage{
rodksinglelab(xlsform)
}
\arguments{
\item{xlsform}{Lastest XLSFORM used to collect data.}
}
\value{

}
\description{
Compile Single Select variable value label into data.frame.
}
\examples{
NOT run
rodksinglelab(xlsform)
}
