% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodkval.R
\name{rodkval}
\alias{rodkval}
\title{Title}
\usage{
rodkval(xlsform, dataName)
}
\arguments{
\item{xlsform}{Latest XLSFORM used to collect the data.}

\item{dataName}{name of the data, which was collected using the XLSFORM
stated above.}
}
\value{

}
\description{
Title
}
