% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodkvar.R
\name{rodkvar}
\alias{rodkvar}
\title{Generates Rscript to add variable labels}
\usage{
rodkvar(xlsform, dataName)
}
\arguments{
\item{xlsform}{Latest XLSFORM used to collect data.}

\item{dataName}{name given to the data.frame collected using the XLSFORM
stated}
}
\value{

}
\description{
Generates Rscript to add variable labels
}
\examples{
Not run
rodkvar(xlsforn,'dataName')
}
