% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodkvartype.R
\name{rodkvartype}
\alias{rodkvartype}
\title{Compiling Variable type into a data.frame from XLSFORM}
\usage{
rodkvartype(xlsform)
}
\arguments{
\item{xlsform}{Latest XLSFORM used to collect the data.}
}
\value{

}
\description{
Compiling Variable type into a data.frame from XLSFORM
}
\examples{
#Not run
rodkvartype(xlsform)

}
