% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabx.R
\name{tabx}
\alias{tabx}
\title{Tabulates single select variables by group.}
\usage{
tabx(row, column, data)
}
\arguments{
\item{row}{Variable to tabulate}

\item{column}{Group variable.}

\item{data}{Data collected using the ODK platform}
}
\value{

}
\description{
Tabulates single select variables by group.
}
\examples{
tabx('row','column','data')
}
