% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toproper.R
\name{toproper}
\alias{toproper}
\title{Converts String text into Proper Format}
\usage{
toproper(x)
}
\arguments{
\item{x}{String or variable to be converted to proper}
}
\value{

}
\description{
Converts String text into Proper Format
}
\examples{
toproper(data$row)
}
