% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlabs.R
\name{varlabs}
\alias{varlabs}
\title{Compile Variable labels into data.frame =}
\usage{
varlabs(xlsform)
}
\arguments{
\item{xlsform}{Lastest XLSForm used to collect data.}
}
\value{

}
\description{
Compile Variable labels into data.frame =
}
\examples{
NOT RUN
varlabs(xlsform)
}
