% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctoimport.R
\name{ctoimport}
\alias{ctoimport}
\title{Fetching Data from SurveyCTO server usin API.}
\usage{
ctoimport(servername, formid, username, password)
}
\arguments{
\item{servername}{SurveyCTO servername}

\item{formid}{Get this from Setting sheet in the XLSForm workbook or in
the form definition in the server.}

\item{username}{SurveyCTO username, without including  the .survecto.com}

\item{password}{SurveyCTO password, it is encouraged not to put the
password in a plain text in your script. Employ it in
 .Renviron files to hide your passwords.}
}
\value{

}
\description{
The function fetch data from surveycto server using formid, which is a form
identifier. It imports data in .csv format, without repeat group data.
}
\examples{
ctoimport('servername','formid','username','password')
}
