% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboimport.R
\name{koboimport}
\alias{koboimport}
\title{Fetching data from KoboToolBox using API (even Humanitarian)}
\usage{
koboimport(servername, formid, username, password)
}
\arguments{
\item{servername}{either 'kobo.humanitarianresponse.info/' or
'kc.kobotoolbox.org' However, the current function is
largely build to stream data from kobotoolbox.org.}

\item{formid}{To download data, a unique identifier of the form is needed
 to construct a URL used to fetch data. The fetch data
 from a given form login to the server and click the form
 and navigate the the form tab. From the link, the text
 between forward slash(/)(after "forms" and before "landing").
 For example, https://kf.kobotoolbox.org/#/forms/aTxwr9Fg4ouTYnRN5tHq2z/landing
formid='aTxwr9Fg4ouTYnRN5tHq2z'}

\item{username}{your username}

\item{password}{you password}
}
\value{

}
\description{
KoBoToolBox provides a suites of tools to collect data in challenging
environment. It is free and opne source and work both online and offline.
KoboTool Box also provides advanced feature and make it useful in
advanced scenarios.It provide data access through RESTAPI, which
automte the workflow of data colelection, processing and visualization.
This is possible even the project is no public, using username and
password authentication.
}
\examples{
###Not run
data<-koboimport(servername,form_id,username,password)

}
