% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redimport.R
\name{redimport}
\alias{redimport}
\title{Fetching data from Project's RedCap server}
\usage{
redimport(token, url)
}
\arguments{
\item{token}{The is unique identifier of projects in RedCap.
This function uses the token to fetch data using API.}

\item{url}{This the API link into the Redcap API. An example is
'https://redcap.xxxxxxxx/api/'}
}
\value{

}
\description{
Fetching data from Project's RedCap server
}
\examples{
###NOT RUN
redimport('token','url')

}
