% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodkval.R
\name{rodkval}
\alias{rodkval}
\title{Generates a Script to add single variable value labels.}
\usage{
rodkval(xlsform, dataName, language = "")
}
\arguments{
\item{xlsform}{Latest XLSFORM used to collect the data.}

\item{dataName}{Use the name used to save the data, in R memory, collected
                using the XLS-form stated above.
stated above.}

\item{language}{Enter the language to be used in value labelling.
If the language is not provided, the most left label
column will be used to generate the value label script.
For example, language="English".}
}
\value{

}
\description{
The function uses expss package to label data.
}
\examples{
NOT run
rodkmult(xlsforn,'dataName','English')
}
