% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabmult.R
\name{tabmult}
\alias{tabmult}
\title{The function generate proportion from a multiple-select variable}
\usage{
tabmult(row, column, data)
}
\arguments{
\item{row}{Multiple-select variable}

\item{column}{Group by variable}

\item{data}{data with multiple select variable}
}
\value{

}
\description{
The function generate proportion from a multiple-select variable
}
\examples{
tabmmult('row','column','data','valuelabel')

}
